+++ src/behavior.rs
@@ -1,5 +1,6 @@
 use egui::{
-    vec2, Color32, Id, Rect, Response, Rgba, Sense, Stroke, TextStyle, Ui, Visuals, WidgetText,
+    vec2, Color32, Id, Rect, Response, Rgba, Sense, Stroke, TextStyle, Ui, Vec2, Visuals,
+    WidgetText,
 };
 
 use super::{ResizeState, SimplificationOptions, Tile, TileId, Tiles, UiResponse};
@@ -61,8 +62,8 @@ pub trait Behavior<Pane> {
 
         // Show a gap when dragged
         if ui.is_rect_visible(rect) && !is_being_dragged {
-            let bg_color = self.tab_bg_color(ui.visuals(), tile_id, active);
-            let stroke = self.tab_outline_stroke(ui.visuals(), tile_id, active);
+            let bg_color = self.tab_bg_color(ui.visuals(), tiles, tile_id, active);
+            let stroke = self.tab_outline_stroke(ui.visuals(), tiles, tile_id, active);
             ui.painter().rect(rect.shrink(0.5), 0.0, bg_color, stroke);
 
             if active {
@@ -74,7 +75,7 @@ pub trait Behavior<Pane> {
                 );
             }
 
-            let text_color = self.tab_text_color(ui.visuals(), tile_id, active);
+            let text_color = self.tab_text_color(ui.visuals(), tiles, tile_id, active);
             ui.painter().galley_with_color(
                 egui::Align2::CENTER_CENTER
                     .align_size_within_rect(galley.size(), rect)
@@ -84,9 +85,18 @@ pub trait Behavior<Pane> {
             );
         }
 
-        self.on_tab_button(tiles, tile_id, &response);
+        self.on_tab_button(tiles, tile_id, response)
+    }
 
-        response
+    /// Show the ui for the tab being dragged.
+    fn drag_ui(&mut self, tiles: &Tiles<Pane>, ui: &mut Ui, tile_id: TileId) {
+        let mut frame = egui::Frame::popup(ui.style());
+        frame.fill = frame.fill.gamma_multiply(0.5); // Make see-through
+        frame.show(ui, |ui| {
+            // TODO(emilk): preview contents?
+            let text = self.tab_title_for_tile(tiles, tile_id);
+            ui.label(text);
+        });
     }
 
     /// Called by the default implementation of [`Self::tab_ui`] for each added button
@@ -94,8 +104,9 @@ pub trait Behavior<Pane> {
         &mut self,
         _tiles: &Tiles<Pane>,
         _tile_id: TileId,
-        _button_response: &Response,
-    ) {
+        button_response: Response,
+    ) -> Response {
+        button_response
     }
 
     /// Return `false` if a given pane should be removed from its parent.
@@ -181,7 +192,13 @@ pub trait Behavior<Pane> {
     }
 
     /// The background color of a tab.
-    fn tab_bg_color(&self, visuals: &Visuals, _tile_id: TileId, active: bool) -> Color32 {
+    fn tab_bg_color(
+        &self,
+        visuals: &Visuals,
+        _tiles: &Tiles<Pane>,
+        _tile_id: TileId,
+        active: bool,
+    ) -> Color32 {
         if active {
             visuals.panel_fill // same as the tab contents
         } else {
@@ -190,7 +207,13 @@ pub trait Behavior<Pane> {
     }
 
     /// Stroke of the outline around a tab title.
-    fn tab_outline_stroke(&self, visuals: &Visuals, _tile_id: TileId, active: bool) -> Stroke {
+    fn tab_outline_stroke(
+        &self,
+        visuals: &Visuals,
+        _tiles: &Tiles<Pane>,
+        _tile_id: TileId,
+        active: bool,
+    ) -> Stroke {
         if active {
             Stroke::new(1.0, visuals.widgets.active.bg_fill)
         } else {
@@ -204,12 +227,28 @@ pub trait Behavior<Pane> {
     }
 
     /// The color of the title text of the tab.
-    fn tab_text_color(&self, visuals: &Visuals, _tile_id: TileId, active: bool) -> Color32 {
+    fn tab_text_color(
+        &self,
+        visuals: &Visuals,
+        _tiles: &Tiles<Pane>,
+        _tile_id: TileId,
+        active: bool,
+    ) -> Color32 {
         if active {
             visuals.widgets.active.text_color()
         } else {
             visuals.widgets.noninteractive.text_color()
         }
+    }
+
+    /// When drag-and-dropping a tile, the candidate area is drawn with this stroke.
+    fn drag_preview_stroke(&self, visuals: &Visuals) -> Stroke {
+        visuals.selection.stroke
+    }
+
+    /// When drag-and-dropping a tile, the candidate area is drawn with this background color.
+    fn drag_preview_color(&self, visuals: &Visuals) -> Color32 {
+        visuals.selection.stroke.color.gamma_multiply(0.5)
     }
 
     /// When drag-and-dropping a tile, how do we preview what is about to happen?
@@ -220,20 +259,15 @@ pub trait Behavior<Pane> {
         parent_rect: Option<Rect>,
         preview_rect: Rect,
     ) {
-        let preview_stroke = visuals.selection.stroke;
-        let preview_color = preview_stroke.color;
+        let preview_stroke = self.drag_preview_stroke(visuals);
+        let preview_color = self.drag_preview_color(visuals);
 
         if let Some(parent_rect) = parent_rect {
             // Show which parent we will be dropped into
             painter.rect_stroke(parent_rect, 1.0, preview_stroke);
         }
 
-        painter.rect(
-            preview_rect,
-            1.0,
-            preview_color.gamma_multiply(0.5),
-            preview_stroke,
-        );
+        painter.rect(preview_rect, 1.0, preview_color, preview_stroke);
     }
 
     /// How many columns should we use for a [`crate::Grid`] put into [`crate::GridLayout::Auto`]?
@@ -246,7 +280,7 @@ pub trait Behavior<Pane> {
     fn grid_auto_column_count(&self, num_visible_children: usize, rect: Rect, gap: f32) -> usize {
         num_columns_heuristic(
             num_visible_children,
-            rect,
+            rect.size(),
             gap,
             self.ideal_tile_aspect_ratio(),
         )
@@ -256,24 +290,35 @@ pub trait Behavior<Pane> {
     fn ideal_tile_aspect_ratio(&self) -> f32 {
         4.0 / 3.0
     }
+
+    // Callbacks:
+
+    /// Called if the user edits the tree somehow, e.g. changes the size of some container,
+    /// clicks a tab, or drags a tile.
+    fn on_edit(&mut self) {}
 }
 
 /// How many columns should we use to fit `n` children in a grid?
-fn num_columns_heuristic(n: usize, rect: Rect, gap: f32, desired_aspect: f32) -> usize {
+fn num_columns_heuristic(n: usize, size: Vec2, gap: f32, desired_aspect: f32) -> usize {
     let mut best_loss = f32::INFINITY;
     let mut best_num_columns = 1;
 
     for ncols in 1..=n {
+        if 4 <= n && ncols == n - 1 {
+            // Don't suggest 7 columns when n=8 - that produces an ugly orphan on a single row.
+            continue;
+        }
+
         let nrows = (n + ncols - 1) / ncols;
 
-        let cell_width = (rect.width() - gap * (ncols as f32 - 1.0)) / (ncols as f32);
-        let cell_height = (rect.height() - gap * (nrows as f32 - 1.0)) / (nrows as f32);
+        let cell_width = (size.x - gap * (ncols as f32 - 1.0)) / (ncols as f32);
+        let cell_height = (size.y - gap * (nrows as f32 - 1.0)) / (nrows as f32);
 
         let cell_aspect = cell_width / cell_height;
         let aspect_diff = (desired_aspect - cell_aspect).abs();
         let num_empty_cells = ncols * nrows - n;
 
-        let loss = aspect_diff + 0.1 * num_empty_cells as f32; // TODO(emilk): weight differently?
+        let loss = aspect_diff * n as f32 + 2.0 * num_empty_cells as f32;
 
         if loss < best_loss {
             best_loss = loss;
@@ -282,4 +327,23 @@ fn num_columns_heuristic(n: usize, rect: Rect, gap: f32, desired_aspect: f32) ->
     }
 
     best_num_columns
+}
+
+#[test]
+fn test_num_columns_heuristic() {
+    // Four tiles should always be in a 1x4, 2x2, or 4x1 grid - NEVER 2x3 or 3x2.
+
+    let n = 4;
+    let gap = 0.0;
+    let ideal_tile_aspect_ratio = 4.0 / 3.0;
+
+    for i in 0..=100 {
+        let size = Vec2::new(100.0, egui::remap(i as f32, 0.0..=100.0, 1.0..=1000.0));
+
+        let ncols = num_columns_heuristic(n, size, gap, ideal_tile_aspect_ratio);
+        assert!(
+            ncols == 1 || ncols == 2 || ncols == 4,
+            "Size {size:?} got {ncols} columns"
+        );
+    }
 }